# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Blocker X - Discord Bot Hosting Platform API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /notifications/{notificationId}/read:
    patch:
      operationId: markNotificationRead
      tags: [notifications]
      summary: Mark a notification as read
      parameters:
        - name: notificationId
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Marked read
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MessageResponse"
...
components:
  schemas:
...
    MessageResponse:
      type: object
      properties:
        message:
          type: string
      required:
        - message
...